% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_heat.R
\name{BmiHeat}
\alias{BmiHeat}
\title{BmiHeat}
\description{
Basic Model Interface (BMI) implementation for the Heat model.
}
\details{
This class implements the Basic Model Interface (BMI) for the 2D Heat
Equation model defined in the Heat R6 class. It provides methods for
initializing the model from a YAML configuration file, advancing the model
in time, and getting/setting model variables. Documentation on the BMI
methods can be found at \url{https://bmi.csdms.io/}.
}
\examples{
config_file <- system.file("extdata", "heat_config.yaml", package = "bmiheatr")
model <- BmiHeat$new()
model$bmi_initialize(config_file)
print(model$get_component_name())
temperature <- model$get_value("plate_surface__temperature")
print(head(temperature))
model$bmi_finalize()
}
\section{Super class}{
\code{\link[bmir:Bmi]{bmir::Bmi}} -> \code{BmiHeat}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BmiHeat-new}{\code{BmiHeat$new()}}
\item \href{#method-BmiHeat-bmi_initialize}{\code{BmiHeat$bmi_initialize()}}
\item \href{#method-BmiHeat-update}{\code{BmiHeat$update()}}
\item \href{#method-BmiHeat-bmi_finalize}{\code{BmiHeat$bmi_finalize()}}
\item \href{#method-BmiHeat-get_component_name}{\code{BmiHeat$get_component_name()}}
\item \href{#method-BmiHeat-get_input_item_count}{\code{BmiHeat$get_input_item_count()}}
\item \href{#method-BmiHeat-get_output_item_count}{\code{BmiHeat$get_output_item_count()}}
\item \href{#method-BmiHeat-get_input_var_names}{\code{BmiHeat$get_input_var_names()}}
\item \href{#method-BmiHeat-get_output_var_names}{\code{BmiHeat$get_output_var_names()}}
\item \href{#method-BmiHeat-get_var_grid}{\code{BmiHeat$get_var_grid()}}
\item \href{#method-BmiHeat-get_var_type}{\code{BmiHeat$get_var_type()}}
\item \href{#method-BmiHeat-get_var_units}{\code{BmiHeat$get_var_units()}}
\item \href{#method-BmiHeat-get_var_itemsize}{\code{BmiHeat$get_var_itemsize()}}
\item \href{#method-BmiHeat-get_var_nbytes}{\code{BmiHeat$get_var_nbytes()}}
\item \href{#method-BmiHeat-get_var_location}{\code{BmiHeat$get_var_location()}}
\item \href{#method-BmiHeat-get_current_time}{\code{BmiHeat$get_current_time()}}
\item \href{#method-BmiHeat-get_start_time}{\code{BmiHeat$get_start_time()}}
\item \href{#method-BmiHeat-get_end_time}{\code{BmiHeat$get_end_time()}}
\item \href{#method-BmiHeat-get_time_units}{\code{BmiHeat$get_time_units()}}
\item \href{#method-BmiHeat-get_time_step}{\code{BmiHeat$get_time_step()}}
\item \href{#method-BmiHeat-get_value}{\code{BmiHeat$get_value()}}
\item \href{#method-BmiHeat-get_value_at_indices}{\code{BmiHeat$get_value_at_indices()}}
\item \href{#method-BmiHeat-set_value}{\code{BmiHeat$set_value()}}
\item \href{#method-BmiHeat-set_value_at_indices}{\code{BmiHeat$set_value_at_indices()}}
\item \href{#method-BmiHeat-get_grid_rank}{\code{BmiHeat$get_grid_rank()}}
\item \href{#method-BmiHeat-get_grid_size}{\code{BmiHeat$get_grid_size()}}
\item \href{#method-BmiHeat-get_grid_type}{\code{BmiHeat$get_grid_type()}}
\item \href{#method-BmiHeat-get_grid_shape}{\code{BmiHeat$get_grid_shape()}}
\item \href{#method-BmiHeat-get_grid_spacing}{\code{BmiHeat$get_grid_spacing()}}
\item \href{#method-BmiHeat-get_grid_origin}{\code{BmiHeat$get_grid_origin()}}
\item \href{#method-BmiHeat-get_grid_node_count}{\code{BmiHeat$get_grid_node_count()}}
\item \href{#method-BmiHeat-get_value_as_matrix}{\code{BmiHeat$get_value_as_matrix()}}
\item \href{#method-BmiHeat-to_matrix}{\code{BmiHeat$to_matrix()}}
\item \href{#method-BmiHeat-display_model}{\code{BmiHeat$display_model()}}
\item \href{#method-BmiHeat-clone}{\code{BmiHeat$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_grid_edge_count"><a href='../../bmir/html/Bmi.html#method-Bmi-get_grid_edge_count'><code>bmir::Bmi$get_grid_edge_count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_grid_edge_nodes"><a href='../../bmir/html/Bmi.html#method-Bmi-get_grid_edge_nodes'><code>bmir::Bmi$get_grid_edge_nodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_grid_face_count"><a href='../../bmir/html/Bmi.html#method-Bmi-get_grid_face_count'><code>bmir::Bmi$get_grid_face_count()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_grid_face_edges"><a href='../../bmir/html/Bmi.html#method-Bmi-get_grid_face_edges'><code>bmir::Bmi$get_grid_face_edges()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_grid_face_nodes"><a href='../../bmir/html/Bmi.html#method-Bmi-get_grid_face_nodes'><code>bmir::Bmi$get_grid_face_nodes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_grid_nodes_per_face"><a href='../../bmir/html/Bmi.html#method-Bmi-get_grid_nodes_per_face'><code>bmir::Bmi$get_grid_nodes_per_face()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_grid_x"><a href='../../bmir/html/Bmi.html#method-Bmi-get_grid_x'><code>bmir::Bmi$get_grid_x()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_grid_y"><a href='../../bmir/html/Bmi.html#method-Bmi-get_grid_y'><code>bmir::Bmi$get_grid_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_grid_z"><a href='../../bmir/html/Bmi.html#method-Bmi-get_grid_z'><code>bmir::Bmi$get_grid_z()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="get_value_ptr"><a href='../../bmir/html/Bmi.html#method-Bmi-get_value_ptr'><code>bmir::Bmi$get_value_ptr()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bmir" data-topic="Bmi" data-id="update_until"><a href='../../bmir/html/Bmi.html#method-Bmi-update_until'><code>bmir::Bmi$update_until()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-new"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-bmi_initialize"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-bmi_initialize}{}}}
\subsection{Method \code{bmi_initialize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$bmi_initialize(config_file = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-update"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-bmi_finalize"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-bmi_finalize}{}}}
\subsection{Method \code{bmi_finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$bmi_finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_component_name"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_component_name}{}}}
\subsection{Method \code{get_component_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_component_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_input_item_count"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_input_item_count}{}}}
\subsection{Method \code{get_input_item_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_input_item_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_output_item_count"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_output_item_count}{}}}
\subsection{Method \code{get_output_item_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_output_item_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_input_var_names"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_input_var_names}{}}}
\subsection{Method \code{get_input_var_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_input_var_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_output_var_names"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_output_var_names}{}}}
\subsection{Method \code{get_output_var_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_output_var_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_var_grid"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_var_grid}{}}}
\subsection{Method \code{get_var_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_var_grid(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_var_type"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_var_type}{}}}
\subsection{Method \code{get_var_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_var_type(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_var_units"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_var_units}{}}}
\subsection{Method \code{get_var_units()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_var_units(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_var_itemsize"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_var_itemsize}{}}}
\subsection{Method \code{get_var_itemsize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_var_itemsize(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_var_nbytes"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_var_nbytes}{}}}
\subsection{Method \code{get_var_nbytes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_var_nbytes(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_var_location"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_var_location}{}}}
\subsection{Method \code{get_var_location()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_var_location(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_current_time"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_current_time}{}}}
\subsection{Method \code{get_current_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_current_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_start_time"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_start_time}{}}}
\subsection{Method \code{get_start_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_start_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_end_time"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_end_time}{}}}
\subsection{Method \code{get_end_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_end_time()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_time_units"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_time_units}{}}}
\subsection{Method \code{get_time_units()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_time_units()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_time_step"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_time_step}{}}}
\subsection{Method \code{get_time_step()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_time_step()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_value}{}}}
\subsection{Method \code{get_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_value(name, dest = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_value_at_indices"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_value_at_indices}{}}}
\subsection{Method \code{get_value_at_indices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_value_at_indices(name, inds, dest = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-set_value"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-set_value}{}}}
\subsection{Method \code{set_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$set_value(name, src)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-set_value_at_indices"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-set_value_at_indices}{}}}
\subsection{Method \code{set_value_at_indices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$set_value_at_indices(name, inds, src)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_grid_rank"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_grid_rank}{}}}
\subsection{Method \code{get_grid_rank()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_grid_rank(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_grid_size"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_grid_size}{}}}
\subsection{Method \code{get_grid_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_grid_size(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_grid_type"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_grid_type}{}}}
\subsection{Method \code{get_grid_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_grid_type(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_grid_shape"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_grid_shape}{}}}
\subsection{Method \code{get_grid_shape()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_grid_shape(grid, shape = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_grid_spacing"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_grid_spacing}{}}}
\subsection{Method \code{get_grid_spacing()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_grid_spacing(grid, spacing = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_grid_origin"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_grid_origin}{}}}
\subsection{Method \code{get_grid_origin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_grid_origin(grid, origin = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_grid_node_count"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_grid_node_count}{}}}
\subsection{Method \code{get_grid_node_count()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_grid_node_count(grid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-get_value_as_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-get_value_as_matrix}{}}}
\subsection{Method \code{get_value_as_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$get_value_as_matrix(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-to_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-to_matrix}{}}}
\subsection{Method \code{to_matrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$to_matrix(src)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-display_model"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-display_model}{}}}
\subsection{Method \code{display_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$display_model()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BmiHeat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BmiHeat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BmiHeat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
