% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{Heat$from_yaml_file}
\alias{Heat$from_yaml_file}
\title{Create a Heat object from a YAML config file (static method)}
\arguments{
\item{yaml_file}{Path to a configuration file in YAML format.}
}
\value{
A new instance of Heat.
}
\description{
Create a Heat object from a YAML config file (static method)
}
\details{
The YAML file should define the keys \code{shape}, \code{spacing}, \code{origin}, and
\code{alpha}.
}
\examples{
# Assuming a YAML file "heat_config.yaml" in the "inst/extdata" directory
# with the following contents:
# shape: [5, 5]
# spacing: [1.0, 1.0]
# origin: [0.0, 0.0]
# alpha: 0.5
cfg_file <- system.file("extdata", "heat_config.yaml", package="bmiheatr")
heat <- Heat$from_yaml_file(cfg_file)
heat$shape
# [1] 5 5
heat$spacing
# [1] 1 1
heat$origin
# [1] 0 0
heat$alpha
# [1] 0.5
}
