% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{solve_2d}
\alias{solve_2d}
\title{solve_2d}
\usage{
solve_2d(temp, spacing, alpha = 1, time_step = 1)
}
\arguments{
\item{temp}{A matrix representing the temperature.}

\item{spacing}{A numeric vector of length 2 (dy, dx) for grid spacing.}

\item{alpha}{Thermal diffusivity (default is 1.0).}

\item{time_step}{Time step (default is 1.0).}
}
\value{
A matrix with the temperatures after time \emph{time_step}.
}
\description{
Heat Equation Model in R
}
\details{
Solve the 2D Heat Equation on a uniform mesh using base R functions.
}
\examples{
z0 <- matrix(0, nrow=5, ncol=5)
z0[3, 3] <- 100.0
solve_2d(z0, spacing=c(1.0, 1.0), alpha=0.25)
#      [,1] [,2] [,3] [,4] [,5]
# [1,]    0  0.0  0.0  0.0    0
# [2,]    0  0.0 12.5  0.0    0
# [3,]    0 12.5 50.0 12.5    0
# [4,]    0  0.0 12.5  0.0    0
# [5,]    0  0.0  0.0  0.0    0
}
